EWPP-1613: Refactoring the contextual filters processing method.
