+++ modules/oe_list_pages_link_list_source/oe_list_pages_link_list_source.services.yml
@@ -5,3 +5,6 @@ services:
   oe_list_pages_link_list_source.contextual_filters_field_mapper:
     class: Drupal\oe_list_pages_link_list_source\ContextualFilterFieldMapper
     arguments: ['@config.factory']
+  oe_list_pages_link_list_source.contextual_filters_values_processor:
+    class: Drupal\oe_list_pages_link_list_source\ContextualFilterValuesProcessor
+    arguments: ['@current_route_match', '@oe_list_pages.list_source.factory', '@oe_list_pages_link_list_source.contextual_filters_builder', '@entity_type.manager', '@oe_list_pages_link_list_source.contextual_filters_field_mapper', '@plugin.manager.multiselect_filter_field']
